Initial support for exchangable VisaLibrary
This commit splits the VisaLibrary object into VisaLibraryBase which contains
wrapper agnostic code and NIVisaLibrary (which inherits from VisaLibraryBase)
and is the ctypes wrapper to the NI-VISA library.

It adds a open_visa_library helper function to be used by the ResourceManager.
The function takes an iterable of strings to specify the libraries to try to open.
The argument format is 'option@wrapper'.

There is only one register wrapper (`ni`) by default. If another <wrapper name> is
provided, it will be looked in a package named `pyvisa-<wrapper name>`. The package
must declare a WRAPPER_CLASS variable that points to a VisaLibraryBase derived class.

Backwards incompatible changes:
- VisaLibrary directly should not be used anymore