Moves list_resources to the library and call it from resource manager
PyVISA was leaking memory because a resource was not properly closed.
Until this commit, `list_resources` was implemented in the ResourceManager
and called `find_resources` and `find_next` in the library.

This design is flawed because different backends might implement this
differently. Moreover, it required all backends to implement
`find_resources` and `find_next` instead of more direct code.

See #135