Fix reseting of line termination for serial
When a serial instruments is set to use the VI_ATTR_TERMCHAR for
VI_ATTR_ASRL_END_IN (default) then it uses the value in VI_ATTR_TERMCHAR.

By default read_termination is None, and VI_ATTR_TERMCHAR ends up being the
default value of LF. But when switching around the read_termination to
something (like CR) then coming back to None will leave VI_ATTR_TERMCHAR
in a different state that at the start (with CR instead of LF) and this will
change the behavior of read on serial instruments.

The above change always makes VI_ATTR_TERMCHAR LF when read_termination is set
to None.