Add expect_termination to read_binary_values (issue 348)
Adding expect_termination argument with read_binary_values and
query_binary_values allows to get back to the behavior of these methods
as it was in pyvisa 1.8 where the very last read termination was not
accounted for. Forced accounting for it makes this method raise a time
out exception in some instruments. See more information in the github
issue 348.